Delbin and msvc improvements.  Warning fixes.

+++ trunk/gpsbabel/jeeps/gpsapp.c
@@ -4301,7 +4301,7 @@ void GPS_D301_Send(UC *data, GPS_PTrack trk, int type)
     p+=sizeof(float);
 
     if (type == 302) {
-      GPS_Util_Put_Float(p,1.0e25);
+      GPS_Util_Put_Float(p, 1.0e25f);
       p+=sizeof(float);
     }
 
@@ -4321,7 +4321,7 @@ void GPS_D304_Send(UC *data, GPS_PTrack trk)
     GPS_Util_Put_Float(p,trk->alt);
     p+=sizeof(float);
 
-    GPS_Util_Put_Float(p, (const float) 1.0e25);
+    GPS_Util_Put_Float(p, 1.0e25f);
     p+=sizeof(float);
 
     /* Not really clear if the members below makes sense to write or not */
